Add AlertingPlugin base class
Add an abstract class that'll
serve as a basis for alerting
plugin representations later.

Upon extra thought, we do not
really need to define handler
class for log collection, but
we could use alerting plugins
themselves as handlers. If we
use this approach, we can get
plugin configs from settings,
store relevant messages there
and call dispatch in the end,
without the need of top-level
AlertingLogHandler instances.