Implement alerting plugin registration
Add the possibility to register plugin
instances that will handle dispatching
of tool run status to monitoring apps.

Since the plugins are essentially just
custom logging handlers, we simply add
them to the root logger. In the end of
the tool run, the dispatch() method is
called on each plugin to send results.