Support including in settings file
Add support for the include key in
the settings file. This will allow
reuse of parts of the same config.

This way, it is possible to define
common alerting settings in 1 file
while only action-specific configs
need to be separate in other files.

The merge_include parameter allows
toggling between overwrite & merge
of matching key in included config.