Do not re-add logging handler multiple times
Check if a handler with given properties was
already added to the root logger, and if so,
do not try to re-add it anymore, in order to
avoid duplicit log messages. This didn't use
to matter that much previously as the logger
handlers were cleaned on every tool restart,
but we have new use cases turning up now and
they will include importing tool's component
to another tool, which could re-init logging
and re-add handlers, causing duplicate logs.