Implement config query tool
Add the QueryTool component that'll enable various operations
over the entity config, such as checking membership of entity
in another entity, or checking if a required security label's
present in an entity as required by another entities' config.

For now, nested membership check is implemented only; labels'
checks are going to be implemented soon in a separate commit.