Add security label processing support
Extend the QueryTool component to add
support of security label processing.
This enables one to define labels (in
the metaparams key) for groups, which
can then be assigned to users allowed
to be members of such groups. The API
of the tool is also extended, so that
the newly added methods are usable by
other scripts importing this library.