Cap change ratio values to one hundred percent
Currently, adding a large count of entities to
a server that already has some entities is not
easily possible since the change ratio can get
determined to be over 100, while the threshold
argument only support values of 1-100. Fix the
situation by capping the possible change ratio
values to a maximum of 100 % in each scenario.